feat(ticket): use native reservation state enum
Persist the fixed hold lifecycle as a PostgreSQL enum and verify the forward Flyway migration. Extract the ticket route family into a native Fastify plugin while retaining root operational composition and existing HTTP behavior.
